requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
mysql-connector-python
passlib[bcrypt]
PyJWT